internal/utils/format/format.go
package format

import (
	"math"
)

// ToDecimal rounds a float to a specified precision
func ToDecimal(input float64, places int) (rounded float64) {
	// If the float is not a number
	if math.IsNaN(input) {
		return math.NaN()
	}

	if places < 0 {
		places = 0
	}

	// Find out the actual sign and correct the input for later
	sign := 1.0
	if input < 0 {
		sign = -1
		input *= -1
	}

	// Use the places arg to get the amount of precision wanted
	precision := math.Pow(10, float64(places))

	// Find the decimal place we are looking to round
	digit := input * precision

	// Get the actual decimal number as a fraction to be compared
	_, decimal := math.Modf(digit)

	// If the decimal is less than .5 we round down otherwise up
	if decimal >= 0.5 {
		rounded = math.Ceil(digit)
	} else {
		rounded = math.Floor(digit)
	}

	// Finally we do the math to actually create a rounded number
	return rounded / precision * sign
}
